new function to process a list of url in one go -alpha

+++ gtfs-netex-test/tool_script_runner.py
@@ -11,6 +11,13 @@ import traceback
 import urllib.request
 from datetime import datetime
 
+import hashlib
+
+def custom_hash(value):
+    sha256_hash = hashlib.sha256()
+    sha256_hash.update(str(value).encode('utf-8'))
+    return sha256_hash.hexdigest()[-8:]
+
 def reversedate():
     # Get the current date
     current_date = datetime.now()
@@ -126,7 +133,7 @@ def download(folder, url, forced=False):
         if filename=="permalink":
             filename="swiss.zip"
         if "?" in filename:  #for data from mobigo, that is fetched by an aspx script with parameters
-            filename="netex.zip"
+            filename="source.zip"
         if forced==False:
             # Download only when not exists
             path=os.path.join(folder, filename)
@@ -163,7 +170,7 @@ def remove_file(path):
     else:
         raise FileNotFoundError(f"File not found: {path}")
 
-def main(script_file,log_file, log_level, todo_block,begin_step):
+def main(script_file,log_file, log_level, todo_block,begin_step,url=None):
     # blockexisted
     blockexisted=False
     # Read the scripts from a file
@@ -172,7 +179,10 @@ def main(script_file,log_file, log_level, todo_block,begin_step):
 
     # go through each block
     for block in data:
-        processdir = processing_data + "/" + block["block"]
+        if url:
+            processdir= processing_data + "/" + todo_block+ "-"+str(custom_hash(url))
+        else:
+            processdir = processing_data + "/" + block["block"]
         blockstop = False
         if not todo_block == block["block"]:
             if not todo_block == "all":
@@ -196,7 +206,10 @@ def main(script_file,log_file, log_level, todo_block,begin_step):
 
             script_name = script['script']
             script_args = script['args']
-            script_download_url = block.get('download_url')
+            if url:
+                script_download_url=url
+            else:
+                script_download_url = block.get('download_url')
             # replace the placeholder for processdir with the correct values and also the other place holders
             script_args = replace_in_string(script_args, "%%dir%%", processdir)
             script_args = replace_in_string(script_args, "%%inputdir%%", input_dir)
@@ -243,6 +256,12 @@ def main(script_file,log_file, log_level, todo_block,begin_step):
                     exit(1)
                 log_all(logging.INFO, f"Command 'download_input_file' executed for url: {script_download_url}\n")
                 continue
+            if script_name == "process_url_list":
+                for url in block.get("download_urls"):
+                    newblock=script_args
+                    main(list_scripts,log_file,log_level,newblock, 0, url=url)
+                # only one process_url_list can be in a block
+                return
             if script_name == 'remove_file':
                 # Execute the download command. The file under the download_url is copied to a folder
                 remove_file(script_input_file_path)
